rag_tools/requirements.txt
langchain==0.1.13
langchain-community==0.0.29
python-dotenv==1.0.1
requests==2.31.0
tqdm==4.66.2
unstructured[all-docs]==0.14.10
web3==6.15.1
PyMuPDF==1.23.8
Pillow==10.1.0
pytesseract==0.3.10